Refactor webcrypto polyfill to use an install function
